Renamed handle-add! and handle-coincidence! as operations on group-networks instead of suboperations of gn:hlt!  (to add-link-to-new-coset! and identify-cosets!, respectively).
Made the policy for selecting which coset of two that participate in a
coincidence explicit, and set it to "remove the one with the higher
number".
- Added tests to catch changes in the number of dead cosets as a 
  consequence of the policy change.
  - One change occurred: B4 moved from 19 to 18 dead cosets.

The change prevents a crash that occurs when trying to build B2xB2.
The latter now builds fine, and has been added to the test suite.